requirements.txt
django==4.1.3
django-environ==0.9.0
django-autoslug==1.9.8
Pillow==9.3.0
